+++ tools/apiview/emitters/cadl-apiview/test/apiview.test.ts
@@ -91,7 +91,7 @@ describe("apiview: tests", () => {
 
   it("describes model", async () => {
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       model Animal {
         species: string;
@@ -115,11 +115,6 @@ describe("apiview: tests", () => {
     }
     `;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       model Animal {
         species: string;
@@ -144,13 +139,13 @@ describe("apiview: tests", () => {
     `;
     const apiview = await apiViewFor(input, {});
     const actual = apiViewText(apiview);
-    compare(expect, actual, 3);
+    compare(expect, actual, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes templated model", async () => {
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       model Thing<T> {
         property: T;
@@ -181,15 +176,10 @@ describe("apiview: tests", () => {
     }
     `;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       model ConstrainedComplex<X extends
         {
-          name: string
+          name: string;
         }
       > {
         prop: X;
@@ -221,13 +211,13 @@ describe("apiview: tests", () => {
     `;
     const apiview = await apiViewFor(input, {});
     const actual = apiViewText(apiview);
-    compare(expect, actual, 3);
+    compare(expect, actual, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes enum", async () => {
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
 
       enum SomeEnum {
@@ -246,11 +236,6 @@ describe("apiview: tests", () => {
       }
     }`;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       enum SomeEnum {
         Plain,
@@ -269,13 +254,13 @@ describe("apiview: tests", () => {
     }`;
     const apiview = await apiViewFor(input, {});
     const actual = apiViewText(apiview);
-    compare(expect, actual, 3);
+    compare(expect, actual, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes union", async () =>{
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       union MyUnion {
         cat: Cat,
@@ -297,11 +282,6 @@ describe("apiview: tests", () => {
       }
     }`;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       model Cat {
         name: string;
@@ -325,13 +305,13 @@ describe("apiview: tests", () => {
     `;
     const apiview = await apiViewFor(input, {});
     const actual = apiViewText(apiview);
-    compare(expect, actual, 3);
+    compare(expect, actual, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes template operation", async () =>{
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       model FooParams {
         a: string;
@@ -357,24 +337,21 @@ describe("apiview: tests", () => {
       >;
     }`;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       op GetFoo is ResourceRead<
         {
           @query
-          name: string,
-          ...FooParams
+          @doc("The name")
+          name: string;
+          ...FooParams;
         },
         {
           parameters:
             {
               @query
-              fooId: string
-            }
+              @doc("The collection id.")
+              fooId: string;
+            };
         }
       >;
 
@@ -390,13 +367,13 @@ describe("apiview: tests", () => {
     }`;
     const apiview = await apiViewFor(input, {});
     const lines = apiViewText(apiview);
-    compare(expect, lines, 3);
+    compare(expect, lines, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes operation with anonymous models", async () =>{
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       op SomeOp(
         param1: {
@@ -408,32 +385,27 @@ describe("apiview: tests", () => {
       ): string;
     }`;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       op SomeOp(
         param1:
           {
-            name: string
+            name: string;
           },
         param2:
           {
-            age: int16
+            age: int16;
           }
       ): string;
     }`;
     const apiview = await apiViewFor(input, {});
     const lines = apiViewText(apiview);
-    compare(expect, lines, 3);
+    compare(expect, lines, 9);
     validateDefinitionIds(apiview);
   });
 
   it("describes interface", async () => {
     const input = `
-    @Cadl.service( { title: "Test" } )
+    @Cadl.service( { title: "Test", version: "1" } )
     namespace Azure.Test {
       interface Foo {
         @get 
@@ -448,11 +420,6 @@ describe("apiview: tests", () => {
     }
     `;
     const expect = `
-    @Cadl.service(
-      {
-        title: "Test"
-      }
-    )
     namespace Azure.Test {
       interface Foo {
         @get
@@ -461,7 +428,6 @@ describe("apiview: tests", () => {
           @path
           name: string
         ): string;
-
         @get
         @route("list")
         list(): string[];
@@ -470,7 +436,40 @@ describe("apiview: tests", () => {
     `;
     const apiview = await apiViewFor(input, {});
     const lines = apiViewText(apiview);
-    compare(expect, lines, 3);
+    compare(expect, lines, 9);
+    validateDefinitionIds(apiview);
+  });
+
+  it("describes string literals", async () => {
+    const input = `
+    @Cadl.service( { title: "Test", version: "1" } )
+    namespace Azure.Test {
+      @doc("Short string")
+      model Foo {};
+
+      @doc("""
+      A long string,
+      with line breaks
+      and stuff...
+      """)
+      model Bar {};
+    }
+    `;
+    const expect = `
+    namespace Azure.Test {
+      @doc("""
+      A long string,
+      with line breaks
+      and stuff...
+      """)
+      model Bar {}
+      @doc("Short string")
+      model Foo {}
+    }
+    `;
+    const apiview = await apiViewFor(input, {});
+    const lines = apiViewText(apiview);
+    compare(expect, lines, 9);
     validateDefinitionIds(apiview);
   });
 });